+++ tests/Unit/Concerns/BitmasksTest.php
@@ -3,6 +3,7 @@
 namespace Henzeb\Enumhancer\Tests\Unit\Concerns;
 
 use Henzeb\Enumhancer\Helpers\Bitmasks\EnumBitmasks;
+use Henzeb\Enumhancer\Tests\Fixtures\BackedEnums\Bitmasks\BitmasksCorrectModifierEnum;
 use Henzeb\Enumhancer\Tests\Fixtures\BackedEnums\Bitmasks\BitmasksIncorrectIntEnum;
 use Henzeb\Enumhancer\Tests\Fixtures\BackedEnums\Bitmasks\BitmasksIntEnum;
 use Henzeb\Enumhancer\Tests\Fixtures\IntBackedEnum;
@@ -11,6 +12,13 @@ use TypeError;
 
 class BitmasksTest extends TestCase
 {
+    protected function setUp(): void
+    {
+        set_error_handler(static function (int $errno, string $errstr): never {
+            throw new TypeError($errstr, $errno);
+        }, E_USER_ERROR);
+    }
+
     public function testShouldReturnCorrectBit(): void
     {
         $this->assertEquals(1, IntBackedEnum::TEST->bit());
@@ -139,5 +147,21 @@ class BitmasksTest extends TestCase
     {
         $this->expectException(TypeError::class);
         EnumBitmasks::getBits(BitmasksIntEnum::class, 64);
+    }
+
+    public function testAllowAsModifier()
+    {
+
+        $this->assertSame(
+            [
+                1 => "Neither",
+                2 => "Read",
+                4 => "Write",
+                6 => "Both",
+            ],
+            BitmasksCorrectModifierEnum::bits()
+        );
+
+        $this->assertEquals(6, BitmasksCorrectModifierEnum::mask(2, 4)->value());
     }
 }